add fix for calendar dates, move ton logic to utils files, add modal for showing errors
