+++ src/components/Calendar.tsx
@@ -5,6 +5,14 @@ import CalendarNumber from '@/components/ui/CalendarNumber';
 import HorizontalButton from '@/components/ui/HorizontalButton';
 import { useUser } from '@/contexts/UserContext';
 import { isUseTon } from '@/hooks/useTelegramStorage';
+import {
+  daysInMonth,
+  generateDates,
+  isFutureDate,
+  getFirstDayOfLastPeriod,
+  predictOvulationAndPeriod,
+  getCalendarNumberType,
+} from '@/utils/periodDates';
 
 type CalendarProps = {
   periodDays: Date[];
@@ -25,11 +33,10 @@ const Calendar: React.FC<CalendarProps> = ({
 }) => {
   const [months, setMonths] = useState<Date[]>([]);
   const [useTon, setUseTon] = useState<boolean>(false);
-
   const { menstruationsLoading } = useUser();
-
   const [changes, setChanges] = useState<{ date: Date; action: 'add' | 'delete' }[]>([]);
   const [firstDayOfLastPeriod, setFirstDayOfLastPeriod] = useState<Date | null>(null);
+  const today = new Date();
 
   useEffect(() => {
     const initialMonths = [];
@@ -52,7 +59,9 @@ const Calendar: React.FC<CalendarProps> = ({
   }, [months]);
 
   useEffect(() => {
-    setFirstDayOfLastPeriod(getFirstDayOfLastPeriod());
+    const fd = getFirstDayOfLastPeriod(periodDays);
+    console.log('First day of last period: ', fd);
+    setFirstDayOfLastPeriod(fd);
   }, [periodDays]);
 
   const loadPreviousMonth = () => {
@@ -65,24 +74,6 @@ const Calendar: React.FC<CalendarProps> = ({
       new Date(prev[prev.length - 1].getFullYear(), prev[prev.length - 1].getMonth() + 1, 1),
     ]);
   };
-
-  const daysInMonth = (date: Date) =>
-    new Date(date.getFullYear(), date.getMonth() + 1, 0).getDate();
-
-  const generateDates = (month: Date) => {
-    const daysInCurrentMonth = daysInMonth(month);
-    const dates = [];
-
-    for (let i = 1; i <= daysInCurrentMonth; i++) {
-      dates.push(new Date(month.getFullYear(), month.getMonth(), i));
-    }
-
-    return dates;
-  };
-
-  const today = new Date();
-
-  const isFutureDate = (date: Date) => date > today;
 
   const handleDateClick = (date: Date) => {
     if (isEditing) {
@@ -106,62 +97,24 @@ const Calendar: React.FC<CalendarProps> = ({
     }
   };
 
-  const getFirstDayOfLastPeriod = () => {
-    if (periodDays.length === 0) return null;
+  const weekDays = ['П', 'В', 'С', 'Ч', 'П', 'С', 'В'];
 
-    const sortedPeriodDays = [...periodDays].sort((a, b) => a.getTime() - b.getTime());
-    let firstDayOfLastPeriod = sortedPeriodDays[sortedPeriodDays.length - 1];
-
-    const maxCheckDays = Math.min(5, sortedPeriodDays.length);
-
-    for (let i = 1; i < maxCheckDays; i++) {
-      const currentDay = sortedPeriodDays[sortedPeriodDays.length - i];
-      const previousDay = sortedPeriodDays[sortedPeriodDays.length - i - 1];
-
-      if ((currentDay.getTime() - previousDay.getTime()) / (1000 * 60 * 60 * 24) >= 2) {
-        break;
-      }
-
-      firstDayOfLastPeriod = previousDay;
-    }
-
-    return firstDayOfLastPeriod;
-  };
-
-  const predictOvulationAndPeriod = (date: Date) => {
-    if (!isFutureDate(date) || !firstDayOfLastPeriod) {
-      return { isPeriod: false, isOvulation: false, isMaybePeriod: false };
-    }
-
-    const daysSinceFirstPeriod = Math.floor(
-      (date.getTime() - new Date(firstDayOfLastPeriod).getTime()) / (1000 * 60 * 60 * 24)
-    );
-
-    const cycleDay = daysSinceFirstPeriod % cycleLength;
-    const isPeriod = cycleDay >= 0 && cycleDay < 5;
-    const isMaybePeriod = cycleDay === 5;
-    const isOvulation = cycleDay >= 12 && cycleDay < 17;
-
-    return { isPeriod, isOvulation, isMaybePeriod };
-  };
+  const monthNames = [
+    'Январь',
+    'Февраль',
+    'Март',
+    'Апрель',
+    'Май',
+    'Июнь',
+    'Июль',
+    'Август',
+    'Сентябрь',
+    'Октябрь',
+    'Ноябрь',
+    'Декабрь',
+  ];
 
   const renderCalendar = () => {
-    const weekDays = ['П', 'В', 'С', 'Ч', 'П', 'С', 'В'];
-    const monthNames = [
-      'Январь',
-      'Февраль',
-      'Март',
-      'Апрель',
-      'Май',
-      'Июнь',
-      'Июль',
-      'Август',
-      'Сентябрь',
-      'Октябрь',
-      'Ноябрь',
-      'Декабрь',
-    ];
-
     return months.map((month, index) => (
       <div key={index} id={`month-${index}`} className="w-full px-4">
         <header className="bg-telegram-bg flex w-full items-center justify-center py-2">
@@ -172,45 +125,17 @@ const Calendar: React.FC<CalendarProps> = ({
 
         <div className="grid w-full grid-cols-7 gap-1">
           {generateDates(month).map((date, index) => {
-            const isCurrentMonth = date.getMonth() === month.getMonth();
-            const isPeriodDay = periodDays.some(
-              (periodDate) => periodDate.toDateString() === date.toDateString()
-            );
-            const { isPeriod, isOvulation, isMaybePeriod } = predictOvulationAndPeriod(date);
-            const isToday = date.toDateString() === today.toDateString();
-            const futurePeriod = isPeriod && isFutureDate(date);
-            const futureOvulation = isOvulation && isFutureDate(date);
-            const futureMaybePeriod = isMaybePeriod && isFutureDate(date);
-            const showPeriod = isPeriodDay || isPeriod;
-            const showOvulation = isOvulation;
-            const showMaybePeriod = isMaybePeriod;
-            const isInChanges = changes.some(
-              (change) => change.date.toDateString() === date.toDateString()
+            const type = getCalendarNumberType(
+              date,
+              periodDays,
+              firstDayOfLastPeriod,
+              cycleLength,
+              today,
+              isEditing,
+              changes
             );
 
-            const getCalendarNumberType = () => {
-              if (isEditing) {
-                if (isPeriodDay) {
-                  return isInChanges ? 'default' : 'periodPast';
-                } else if (isInChanges) {
-                  return 'select';
-                }
-                return 'default';
-              }
-              if (showPeriod) {
-                return futurePeriod ? 'periodFuture' : 'periodPast';
-              }
-              if (showMaybePeriod) {
-                return 'select';
-              }
-              if (showOvulation) {
-                return futureOvulation ? 'ovulationFuture' : 'ovulationPast';
-              }
-              if (isToday) {
-                return 'today';
-              }
-              return 'default';
-            };
+            const isCurrentMonth = date.getMonth() === month.getMonth();
 
             return (
               <div key={index} className="relative flex flex-col items-center">
@@ -220,7 +145,7 @@ const Calendar: React.FC<CalendarProps> = ({
                   }`}
                   onClick={() => handleDateClick(date)}>
                   <CalendarNumber
-                    type={getCalendarNumberType()}
+                    type={type}
                     number={date.getDate()}
                     className={!isCurrentMonth ? 'text-gray-400' : ''}
                   />
@@ -248,7 +173,7 @@ const Calendar: React.FC<CalendarProps> = ({
       <div className="fixed left-0 top-0 z-10 h-[56px] w-full bg-gradient-to-b from-white to-[#EEF6F8]" />
       <div className="fixed top-0 z-20 flex h-[56px] w-full items-center justify-between px-4">
         <div className="grid w-full grid-cols-7 gap-1">
-          {['П', 'В', 'С', 'Ч', 'П', 'С', 'В'].map((day, index) => (
+          {weekDays.map((day, index) => (
             <div key={index} className="flex flex-col items-center">
               <span className="text-[12px] font-semibold text-text-dark opacity-60">{day}</span>
             </div>
